ai-service/requirements.txt
# Core & Environment
python-dotenv>=1.0.0
groq>=0.5.0
numpy>=1.24.0,<2.0.0

# Web Server
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9

# LangChain & AI
langchain-chroma>=0.1.0        # LangChain wrapper for Chroma vector store
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.1.0
sentence-transformers>=2.6.0

# Vector Database
chromadb>=0.4.0

# Database Connectivity
mysql-connector-python>=8.0.0  # To fetch products from SQL

# Audio
pyttsx3>=2.90                  # Text-to-Speech (offline)
pydub>=0.25.1                  # Audio processing